MAINT: rename metadata folder into config
