+++ imageio/config/__init__.py
@@ -1,0 +1,7 @@
+from .extensions import known_extensions
+from .plugins import known_plugins
+
+__all__ = [
+    "known_plugins",
+    "known_extensions"
+]
