+++ imageio/config/extensions.py
